Initial project content
This is the initial commit of the project's content.  It currently
builds a Gem properly and should be ready to publish as needed.  In
addition it defines all of the bash scripts that add functionalty to
the git command.  These scripts delegate to command types, which are
largely unimplemented at the moment.  The api-key and project-id
configuration properties can be retrieved and are prompted for if they
have not been set.