Prompt for project
Previously, if no project id was associated with a Git repository, the
user would be prompted to enter a project id.  The easiest way to get
this id was to parse a URI from a web browser and this wasn't idea.
This change now prompts the user for a project id by providing a
complete listing of all projects that a user is a member of and
allowing a selection from the list.