Prompt for story and branch name
Previously, calling git start did nothing.  This change adds the
beginnging start behavior.  git start can now be called in three ways:

1. with a story id, to start a specific story
2. with a story type, to be prompted for with a list of stories of
   that type
3. with nothing, to be prompted with a list of stories of all types

Once a story has been selected in one of these three ways, the user is
prompted for a branch name.  The branch name is then prepended with
the story id to generate the full branch name that will be used.