Create branch
Previously, when starting a story nothing was done in your local git
repository to help you with branches and commits.  Now, when you start
a story, a branch is created based on the name that you input earlier.
In addition, a prepare-commit-msg commit hook that ensures that the
story id is attached is added to the repository.