Do not add story id to amended commit
Previously, when a commit was amended, the prepare-commit-msg would
hook would add the story id.  Typically, the existing commit message
already had that story id so this would cause it to show twice. This
change updates the commit hook so that it will not add the story id to
any commit based on another commit.

[Closes #4]