No complete
Previously, any time finish was called, the commit message would
include the Pivotal Tracker keyword 'Completes' in it.  This would
trigger a state change in tracker.  However, in the case of work that
spanned repositories, a branch of work could be complete before the
issue was actually complete.  This change allows a user to pass in a
--no-complete option which will supress the 'Completes' keyword.